requirements.txt
backports.csv==1.0.7
certifi==2019.3.9
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.1
django-crispy-forms==1.7.2
django-filter==2.1.0
django-formtools==2.1
django-import-export==1.2.0
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
et-xmlfile==1.0.1
future==0.17.1
httplib2==0.12.3
idna==2.8
itypes==1.1.0
jdcal==1.4.1
Jinja2==2.10.1
MarkupSafe==1.1.1
mysqlclient==1.4.2
odfpy==1.4.0
openpyxl==2.6.2
Pillow==6.0.0
PyJWT==1.7.1
pytz==2019.1
PyYAML==5.1
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
tablib==0.13.0
uritemplate==3.0.0
urllib3==1.25.3
xlrd==1.2.0
xlwt==1.3.0
